Switched mouseSignal stuff to just use doubles instead of int since that's what the JS is actually producing.

+++ signalTest.cpp
@@ -29,7 +29,7 @@ class TestObj {
         cout << "TestObj::signalAcceptorTestMethod(): " << s_ << endl;
     }
 
-    void mouseAcceptorTestMethod(const std::pair<int, int> &mouseLocation) {
+    void mouseAcceptorTestMethod(const std::pair<double, double> &mouseLocation) {
         
         cout << "TestObj::mouseAcceptorTestMethod(): x = " << mouseLocation.first << ", y = " << mouseLocation.second << endl;
     }
@@ -117,16 +117,16 @@ int main() {
     auto svg = cl2::SVG("svg1", 400, 300, getStrId(), generatedDiv.domElement_);
     svg.setAttributes({{"viewBox", val("20 20 60 60")}, {"style", val("border: 1px solid black")}});
 
-    auto mouseSignal = make_shared<MouseSignal<std::pair<int, int>>>(svg, "click");
+    auto mouseSignal = make_shared<MouseSignal<std::pair<double, double>>>(svg, "click");
 
-auto svgMouseClickAcceptor = make_shared<ObjectAcceptor<std::pair<int, int>, TestObj>>(tobjSptr);
+auto svgMouseClickAcceptor = make_shared<ObjectAcceptor<std::pair<double, double>, TestObj>>(tobjSptr);
     svgMouseClickAcceptor->setSignalAcceptorMethod(&TestObj::mouseAcceptorTestMethod);
 
 
     // auto svgMouseClickOutput =
-    //     sb.textInputWSS<std::pair<int, int>>("svgMouseClickOutput", "SVG Mouse click location.");
+    //     sb.textInputWSS<std::pair<double, double>>("svgMouseClickOutput", "SVG Mouse click location.");
 
-    sb.connect<std::pair<int, int>>(mouseSignal, svgMouseClickAcceptor);
+    sb.connect<std::pair<double, double>>(mouseSignal, svgMouseClickAcceptor);
 
     const auto circle1 = cl2::WebElement("circle", "circle1", getStrId(), svg.domElement_);
     circle1.domElement_.call<void>("setAttribute", val("r"), val(45));
